Add data processing script for iPhone Reddit comments
This script creates a data processing pipeline for Reddit comments about the iPhone. First, we define the data structure of each comment. Next, we identify all posts related to the iPhone and remove any special characters in the comments. Finally, we perform some topic modeling.

This is an essential part of our analysis as it allows us to understand the main topics being discussed in the comments.

Additionally, add relevant stopwords and terms related to technology and common terms for better results in topic modeling. This further refines our analysis and provides more accurate results.